+++ bluesky/config/defaults.py
@@ -92,32 +92,29 @@ _DEFAULTS = {
     },
     "ecoregion": {
         "lookup_implementation": "ogr",
+        "try_nearby_on_failure": False,
         "skip_failures": True,
         "default": None
     },
     "fuelbeds": {
-        # The following defaults are defined in the fccsmap package,
-        # so they could be removed from here
-        "fccs_version": "2",
-        "ignored_percent_resampling_threshold": 99.9,
-        "ignored_fuelbeds": ['0', '900'],
-        "no_sampling": False,
-        "use_all_grid_cells": False,
-        "sampling_radius_factors": None,
         "skip_failures": False,
 
-        # The following defaults are defined in the fccsmap package
-        # and are based on the location of the package in the file
-        # system. So, let fccsmap set defaults
-        # "fccs_fuelload_file": None,
-        # "fccs_fuelload_param": None,
-        # "fccs_fuelload_grid_resolution": None,
+        # The following defaults are defined in the fccsmap package,
+        # so they could be removed from here
+        "ignored_fuelbeds": ['0', '900'],
+        "ignored_percent_resampling_threshold": 99.9,
+        "insignificance_threshold": 10.0,
+        "max_fuelbed_count_threshold": 5,
+        "no_sampling": False,
+        "sampling_radius_km": 1.0,
+        "sampling_radius_factors": None,
+        "use_all_grid_cells": False,
 
-        # The following defaults are *not* defined in fccsmap package
-        # If user override truncation threshold defaults by setting to
-        # None or 0, we won't truncate by that criterion
-        "truncation_percentage_threshold": 90.0,
-        "truncation_count_threshold": 5,
+        # fccs_version only comes into play if files aren't specified
+        "fccs_fuelload_tile_sets": [], # each fileset is a dict with 'directory' and optionally 'index_shapefile'
+        "fccs_fuelload_files": [],
+        "fccs_version": "2",
+
         # Allow summed fuel percentages to be between 99.5% and 100.5%
         "total_pct_threshold": 0.5
     },
@@ -141,6 +138,7 @@ _DEFAULTS = {
             "enabled": False,
             "run_consume_on_failure": True
         },
+        "summarize_fuel_loadings": False,
         "consume_settings": {
             # TODO: Confirm with Susan P, Susan O. to confirm that these
             #    burn-type specific settings (defaults and synonyms) are
@@ -404,7 +402,7 @@ _DEFAULTS = {
         "delete_working_dir_if_no_error": True,
         "handle_existing": "fail",
         "hysplit": {
-            "binary": 'hyts_std',
+            "binary": 'hyts_std-v5.2.3',
             "start_hours": [0],
             "heights": [10, 100, 1000],
             "vertical_motion": 0, # 0 = from met file
@@ -454,6 +452,11 @@ _DEFAULTS = {
             # }
         },
         "hysplit": {
+            "emissions_split": {
+                "enabled": False,
+                "target_pm25": 10.0,  # (specified in ug/m^3)
+                "min_pm25": None  # (specified in ug/m^3)
+            },
             "binaries": {},
             "skip_invalid_fires": False,
 
@@ -777,7 +780,7 @@ _DEFAULTS = {
             # Period's initial vertical dispersion at the surface (m)
             "OZINTA": 0.0,
 
-            # Period's background PM (ug/m3)
+            # Period's background PM (ug/m^3)
             "BKGPMA": 0.0,
 
             # Period's background CO (ppm)