+++ bluesky/consumption/consume/consumeutils.py
@@ -107,8 +107,14 @@ def _get_setting(location, field):
 def _apply_settings(fc, location, burn_type, fire_type):
     # Read settings here instead of at module scope to support unit testing
 
+    # 'wf' gets translated to 'wildfire' in the fire model, model, but the
+    # consume_settings configuration uses 'wf'
+    if fire_type == 'wildfire':
+        fire_type = 'wf'
+
     valid_settings = dict(SettingsManager().SETTINGS[burn_type],
         **SettingsManager().SETTINGS['all'])
+
     for field, d in valid_settings.items():
         value = _get_setting(location, field)